Modifying an Existing Rails Application for OpenShift
In this asciicast we will modify an existing Rails application. For this, we will use Hack The Budget ( https://github.com/sdhull/hack_the_budget ) , a Rails + Mongo application recently used for a hackathon that OpenShift participated in.

To start, we've checked out the code from Github and created an OpenShift ( http://openshift.redhat.com ) Ruby application with an embedded Mongo database.

Just a note, at the end, we stopped the process because the deploy can take a few minutes the first time (it needs time to install your gems).